+++ src/pages/projects/Sections.js
@@ -6,12 +6,13 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 
 const useStyles = makeStyles((theme) => ({
     section: {
-        position: 'relative'
+        position: 'relative',
+        paddingTop: '80px!important'
     },
     sectionIndex: {
         [theme.breakpoints.up('lg')]: {
             position: 'absolute',
-            top: 0,
+            top: '50px',
             left: 0,
             transform: 'translateX(-100%)'
         }
@@ -25,7 +26,7 @@ const Sections = ({ classes, ...props }) => {
 
     return (
         <>
-            <Grid id="brief" container className={internalClasses.section} component="section" item xs={12}>
+            <Grid id="brief" spacing={4} container className={internalClasses.section} component="section" item xs={12}>
                 <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
                     <Typography variant="body1" className="fs-biggest fw-800">
                         01.
@@ -36,12 +37,32 @@ const Sections = ({ classes, ...props }) => {
                         {project?.brief}
                     </Typography>
                 </Grid>
+                <Grid item xs={12} md={4}>
+                    <div style={{ height: '60vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
+                <Grid item xs={12} md={4}>
+                    <div style={{ height: '60vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
+                <Grid item xs={12} md={4}>
+                    <div style={{ height: '60vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
             </Grid>
-            <Grid id="research" container className={internalClasses.section} component="section" item xs={12}>
+            <Grid
+                id="research"
+                spacing={4}
+                container
+                className={internalClasses.section}
+                component="section"
+                item
+                xs={12}
+            >
                 <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
                     <Typography variant="body1" className="fs-biggest fw-800">
                         02.
                     </Typography>
+                </Grid>
+                <Grid item xs={12}>
+                    <div style={{ height: '60vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
                 </Grid>
                 <Grid item xs={12} sm={9} lg={12}>
                     <Typography variant="body1" className={clsx(internalClasses.sectionContent, 'fs-200')}>
@@ -49,7 +70,15 @@ const Sections = ({ classes, ...props }) => {
                     </Typography>
                 </Grid>
             </Grid>
-            <Grid id="visualConcept" container className={internalClasses.section} component="section" item xs={12}>
+            <Grid
+                id="visualConcept"
+                spacing={4}
+                container
+                className={internalClasses.section}
+                component="section"
+                item
+                xs={12}
+            >
                 <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
                     <Typography variant="body1" className="fs-biggest fw-800">
                         03.
@@ -60,8 +89,28 @@ const Sections = ({ classes, ...props }) => {
                         {project?.visualConcept}
                     </Typography>
                 </Grid>
+                <Grid item xs={12} md={6}>
+                    <div style={{ height: '40vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
+                <Grid item xs={12} md={6}>
+                    <div style={{ height: '40vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
+                <Grid item xs={12} md={6}>
+                    <div style={{ height: '40vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
+                <Grid item xs={12} md={6}>
+                    <div style={{ height: '40vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                </Grid>
             </Grid>
-            <Grid id="prototype" container className={internalClasses.section} component="section" item xs={12}>
+            <Grid
+                id="prototype"
+                spacing={4}
+                container
+                className={internalClasses.section}
+                component="section"
+                item
+                xs={12}
+            >
                 <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
                     <Typography variant="body1" className="fs-biggest fw-800">
                         04.
@@ -71,6 +120,11 @@ const Sections = ({ classes, ...props }) => {
                     <Typography variant="body1" className={clsx(internalClasses.sectionContent, 'fs-200')}>
                         {project?.prototype}
                     </Typography>
+                </Grid>
+                <Grid container item xs={12} justifyContent="center">
+                    <Grid item xs={12} md={5}>
+                        <div style={{ height: '80vh', width: '100%', backgroundColor: '#C4C4C4' }}></div>
+                    </Grid>
                 </Grid>
             </Grid>
         </>